Run lastBlockJob only once during startup.
Disable both the background periodic job and the call from blocknotify.

It was not buggy per-se but it was misleading.  The periodic lastBlockJob would
successfully handle ordinary transactions and reorgs but not double-spends.  It
might erroneously lead Baron deployments to think it is working with a broken
walletnotify configuration.

Keep its plumbing because we may decide to bring it back later.